requirements.txt
fastapi
uvicorn
anthropic
google-genai
python-dotenv
httpx
